requirements_simple.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
Pillow==10.0.1
python-decouple==3.8
celery==5.3.4
redis==5.0.1
django-extensions==3.2.3
django-storages==1.14.2
boto3==1.29.7
reportlab==4.0.4
openpyxl==3.1.2
python-dateutil==2.8.2
pytz==2023.3
django-allauth==0.57.0
djangorestframework-simplejwt==5.3.0
django-phonenumber-field==7.1.0
phonenumbers==8.13.19
django-crispy-forms==2.1
crispy-bootstrap5==0.7
requests==2.31.0
drf-yasg==1.21.7
